Move bulk of code out of __init__ file
